refactor: Re-organize Metrics into clearer structure
