+++ Sources/UnleashProxyClientSwift/Utils/Date+iso8601String.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+extension Date {
+    func iso8601String() -> String {
+        ISO8601DateFormatter().string(from: self)
+    }
+}
